Split hooks into their own modules, regroup api types with api hooks
