+++ src/api/helpers.tsx
@@ -1,16 +1,14 @@
-const clientId = '9031tpad4quocef3dgtu0bbnbupmdr';
+import { TWITCH_CLIENT_ID } from '../constants';
 
-const getOauthUrl = () => {
-  const params = new URLSearchParams({
-    client_id: clientId,
-    redirect_uri: 'http://localhost:3000',
-    response_type: 'token',
-    scope: ['user:read:email', 'channel:read:subscriptions'].join(' '),
-  });
+const oauthParams = new URLSearchParams({
+  client_id: TWITCH_CLIENT_ID,
+  redirect_uri: 'http://localhost:3000',
+  response_type: 'token',
+  scope: ['user:read:email', 'channel:read:subscriptions'].join(' '),
+});
 
-  return `https://id.twitch.tv/oauth2/authorize?${params}`;
-};
+const oauthUrl = `https://id.twitch.tv/oauth2/authorize?${oauthParams}`;
 
 export const redirectForToken = (): void => {
-  window.location.replace(getOauthUrl());
+  window.location.replace(oauthUrl);
 };
