+++ src/components/atoms/index.tsx
@@ -1,5 +1,5 @@
 export { FollowerCount } from './FollowerCount';
 export { LatestFollower } from './LatestFollower';
 export { LatestSubscriber } from './LatestSubscriber';
-export { UserDisplayName } from './UserDisplayName';
 export { SubscriberCount } from './SubscriberCount';
+export { UserDisplayName } from './UserDisplayName';
