+++ src/constants.tsx
@@ -1,1 +1,2 @@
 export const TWITCH_API_ENDPOINT = 'https://api.twitch.tv/helix/';
+export const TWITCH_CLIENT_ID = '9031tpad4quocef3dgtu0bbnbupmdr';
