+++ src/hooks/useTwitchCurrentUser.ts
@@ -1,0 +1,6 @@
+import { useTwitchUsers } from './useTwitchUsers';
+
+export const useTwitchCurrentUser = () => {
+  const { data } = useTwitchUsers();
+  return data?.data[0];
+};
